+++ recipes/agent_windows.rb
@@ -1,3 +1,0 @@
-# Installs and configures the New Relic Infrastructure agent on Windows
-
-raise 'Windows chef support for New Relic Infrastructure agent not yet available'